requirements.txt
openai
python-dotenv
spider-client
websockets
asyncio
requests
pdfminer.six
bs4
fastapi
python-multipart
uvicorn